requirements.txt
click==8.0.1
colorama==0.4.4    
Flask==2.0.1       
itsdangerous==2.0.1
Jinja2==3.0.1      
MarkupSafe==2.0.1  
numpy==1.20.3      
Pillow==8.2.0      
scipy==1.6.3       
Werkzeug==2.0.1